% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manhattanr.R
\name{manhattanr}
\alias{manhattanr}
\title{Creates a manhattanr object}
\source{
The pre-processing is mostly the same as the
  \code{\link[qqman]{manhattan}} function from the
  \href{https://github.com/stephenturner/qqman}{\code{qqman}} package by
  \href{http://www.gettinggeneticsdone.com/}{Stephen Turner}
}
\usage{
manhattanr(x, chr = "CHR", bp = "BP", p = "P", snp = "SNP",
  gene = "GENE", logp = TRUE, ...)
}
\arguments{
\item{x}{A data.frame with columns "BP," "CHR," "P," and optionally, "SNP."}

\item{chr}{A string denoting the column name for the chromosome. Defaults to
PLINK's "CHR." Said column must be numeric. If you have X, Y, or MT
chromosomes, be sure to renumber these 23, 24, 25, etc.}

\item{bp}{A string denoting the column name for the chromosomal position.
Defaults to PLINK's "BP." Said column must be numeric.}

\item{p}{A string denoting the column name for the p-value. Defaults to
PLINK's "P." Said column must be numeric.}

\item{snp}{A string denoting the column name for the SNP name (rs number).
Defaults to PLINK's "SNP." Said column should be a character. This
argument is optional}

\item{gene}{A string denoting the column name for the GENE name. Defaults to
PLINK's "SNP." Said column should be a character. This argument is optional}

\item{logp}{If TRUE, the -log10 of the p-value is plotted. It isn't very
useful to plot raw p-values, but plotting the raw value could be useful for
other genome-wide plots, for example, peak heights, bayes factors, test
statistics, other "scores," etc.}
}
\description{
An object of class manhattanr includes all the needed information for
producing a manhattan plot. In the same spirit as the \link{heatmaply}
package, the goal is to seperate the pre-processing of the manhattan plot
elements from the graphical rendaring of the object, which could be done
using any graphical device including \code{\link[plotly]{plot_ly}} and
\code{\link[graphics]{plot}} in base \code{R}.
}
\seealso{
\code{\link[qqman]{manhattan}},
  \url{https://github.com/stephenturner/qqman},
   \href{https://github.com/nstrayer/D3ManhattanPlots}{D3ManhattanPlots}
}

