% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/significantSNP.R
\docType{data}
\name{significantSNP}
\alias{significantSNP}
\title{Character vector of SNPs to highlight}
\format{A character vector with 21 elements}
\usage{
significantSNP
}
\description{
SNPs from \code{\link{HapMap}} dataset at with \code{p-value < 1e-6}
}
\seealso{
\code{\link{HapMap}}
}
\keyword{datasets}

