% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/significantSNP-data.R
\docType{data}
\name{veteran}
\alias{veteran}
\title{Character vector of SNPs to highlight}
\format{A character vector with 21 elements}
\usage{
veteran
}
\description{
SNPs from \code{hapmap} dataset at with \code{p-value < 1e-6}
}
\seealso{
\code{hapmap}
}
\keyword{datasets}

