% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manhattanly.R
\name{manhattanly}
\alias{manhattanly}
\alias{manhattanly.default}
\alias{manhattanly.manhattanr}
\title{Creates a plotly manhattan plot}
\usage{
manhattanly(x, col = RColorBrewer::brewer.pal(n = 9, name = "Set1"),
  labelChr = NULL, suggestiveline = -log10(1e-05),
  genomewideline = -log10(5e-08), suggestiveline_color = "blue",
  genomewideline_color = "red", highlight = NULL, showlegend = TRUE,
  showgrid = TRUE, xlab = NULL, ylab = "-log10(p)",
  title = "Manhattan Plot", ...)
}
\arguments{
\item{col}{A character vector indicating which colors to alternate.}

\item{labelChr}{A character vector equal to the number of chromosomes
specifying the chromosome labels (e.g., \code{c(1:22, "X", "Y", "MT")}).}

\item{suggestiveline}{Where to draw a "suggestive" line. Default
-log10(1e-5). Set to FALSE to disable.}

\item{genomewideline}{Where to draw a "genome-wide sigificant" line. Default
-log10(5e-8). Set to FALSE to disable.}

\item{suggestiveline_color}{color of "suggestive" line. Only used if
\code{suggestiveline} is not set to \code{FALSE}}

\item{genomewideline_color}{color of "genome-wide sigificant" line. Only used if
\code{genomewideline} is not set to \code{FALSE}}

\item{highlight}{A character vector of SNPs in your dataset to highlight.
These SNPs should all be in your dataset.}

\item{showlegend}{Should a legend be shown. Default is \code{TRUE}.}

\item{...}{other parameters passed to \link{manhattanr}}
}
\value{
A manhattan plot.
}
\description{
Creates a manhattan plot from PLINK assoc output (or any data frame with
chromosome, position, and p-value).
}
\examples{
\dontrun{
library(qqman)
manhattanly(gwasResults)
}
}

