% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/significantSNP-data.R
\docType{data}
\name{significantSNP}
\alias{significantSNP}
\title{Character vector of SNPs to highlight}
\format{A character vector with 21 elements}
\usage{
significantSNP
}
\description{
SNP rs identifiers from \code{hapmap} dataset that are significant at \code{p-value
< 1e-6}
}
\seealso{
\code{hapmap}
}
\keyword{datasets}

