% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/significantSNP-data.R
\docType{data}
\name{significantSNP}
\alias{significantSNP}
\title{Character vector of SNPs to highlight}
\format{A character vector with 20 elements}
\usage{
significantSNP
}
\description{
SNP rs identifiers from \code{\link{HapMap}} dataset that are significant at \code{p-value
< 1e-6}
}
\seealso{
\code{\link{HapMap}}
}
\keyword{datasets}

